Refactor project structure to separate infrastructure components
- Created a new module for infrastructure persistence, moving SQL models, Redis, MongoDB, and PostgreSQL related code to `infrastructure.persistence`.
- Updated import paths across the application to reflect the new structure.
- Removed the old models initialization file as it is no longer needed.
- Ensured all services and routes are correctly referencing the new import paths for models and database connections.
