+++ src/V_APP/Data_Module/routes/events.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 from typing import List
-from models.pydantic_models import EventResponse
+from application.schemas import EventResponse
 from services.event_service import get_events
 
 router = APIRouter()
