+++ src/V_APP/Data_Module/routes/notifications.py
@@ -10,7 +10,7 @@ from fastapi import APIRouter, HTTPException, Query
 from pydantic import BaseModel, ConfigDict
 from bson import ObjectId
 
-from db.mongo import notifications_collection
+from infrastructure.persistence.mongo import notifications_collection
 from services.notification_service import create_notification  # re-export for convenience
 
 __all__ = ["router", "create_notification"]
