+++ src/V_APP/Data_Module/services/statistics_service.py
@@ -13,13 +13,13 @@ from typing import Dict, Any, List, Optional
 from datetime import datetime, timezone, timedelta
 from bson import ObjectId
 
-from db.mongo import (
+from infrastructure.persistence.mongo import (
     agent_detections_collection,
     decision_events_collection,
     statistics_hourly_collection,
     get_hourly_statistics
 )
-from db.redis import (
+from infrastructure.persistence.redis import (
     get_counter,
     get_counter_range,
     get_all_active_counters,
